nett/netmon: add netmon hook for os network change notifications
Update tailscale/corp#20677

While hunting down DNS issues, it was observed that
we often attempt to rebind to interfaces before they're
completely configured by the OS (notably on macOS and iOS).
The AF_ROUTE messages aren't lying, they're just not
taking into account other setup that has to happen,
notably the system setting the interface nameservers.
When we react based on the route updates, it's often
too quickly.

This replumbs netmon so that (optionally) rather than reacting to every
system message by checking for interesting changes, we
use that only to update our state, and recheck for interesting
network changes when we get a poke from the upper layers
indicating that a network change has happened.  Notably on
Apple, NWPathMonitor seems to call us back at the 'right'
time to check for potentially interesting interface changes.

This seems to reduce thrashing during network
transitions, and avoids some of the bad states we were
previously ending up in.

This is entirely opt in during the construction of the netmon
and will have no effect until/unless a specific client chooses
to use it.

Signed-off-by: Jonathan Nobels <jonathan@tailscale.com>